Refactor: Change uint256 to uint in various contracts for consistency and optimization
- Updated data types from uint256 to uint in GmxVenueValidator, MockVenueValidator, and related interfaces for improved gas efficiency.
- Simplified constructor definitions in Dictatorship, FeeMarketplace, FeeMarketplaceStore, PuppetToken, PuppetVoteToken, RouterProxy, and other contracts by removing unnecessary line breaks.
- Enhanced readability by consolidating function signatures and removing redundant line breaks in various contracts.
- Adjusted error handling and revert messages for better clarity in error scenarios.

+++ src/position/interface/IGmxTypes.sol
@@ -35,14 +35,14 @@ interface IBaseOrderUtils {
     }
 
     struct CreateOrderParamsNumbers {
-        uint256 sizeDeltaUsd;
-        uint256 initialCollateralDeltaAmount;
-        uint256 triggerPrice;
-        uint256 acceptablePrice;
-        uint256 executionFee;
-        uint256 callbackGasLimit;
-        uint256 minOutputAmount;
-        uint256 validFromTime;
+        uint sizeDeltaUsd;
+        uint initialCollateralDeltaAmount;
+        uint triggerPrice;
+        uint acceptablePrice;
+        uint executionFee;
+        uint callbackGasLimit;
+        uint minOutputAmount;
+        uint validFromTime;
     }
 
     struct CreateOrderParams {
